requirements.txt
openai
tiktoken
requests
streamlit
pillow